grafana4: dashboard: do not override the title.
grafana4_dashboard.present() was liable to override the user-provided dashboard
title with its name (which is actually the slug, as hinted in the documentation
of grafana4.get_dashboard()). It makes sense for grafana4_dashboard.present()
to enforce a title to fulfill Grafana's requirements, but the slug (normalized,
non-numerical, URL-friendly identifier) should be used as a title
(human-friendly identifier) only as a last resort.